move out safety and greetings from model wrapper and into dialogue_object wrapper
fix call to dm.step

fix add preprocessed chat to memory
